+++ appinfo/info.xml
@@ -45,7 +45,7 @@
     <screenshot>https://raw.githubusercontent.com/nextcloud/forms/master/screenshots/forms3.png</screenshot>
 
     <dependencies>
-        <nextcloud min-version="19" max-version="22" />
+        <nextcloud min-version="20" max-version="22" />
     </dependencies>
 
 	<navigations>
