+++ css/forms.scss
@@ -22,7 +22,6 @@
 
 // Various variables used by this app
 :root {
-	--header-height: 50px; // TODO Remove this line, when minversion is set to 20
 	--top-bar-height: 60px;
 }
